+++ src/NDItk/base/IntegerListRecord.hpp
@@ -24,45 +24,7 @@ class IntegerListRecord : protected ListRecord< IntegerListRecord, int > {
 
 protected:
 
-  /**
-   *  @brief Write the record content
-   *
-   *  This assumes that the content is defined.
-   *
-   *  @param[in] iter   the current position in the output
-   */
-  template< typename OutputIterator >
-  void write( OutputIterator& iter ) const {
-
-    std::ostringstream buffer;
-
-    auto full = this->data().size() / 5;
-    auto partial = this->data().size() - full * 5;
-    auto x = this->data().begin();
-
-    while ( full-- ) {
-
-      buffer << "    " << x[0];
-      buffer << ' ' << x[1];
-      buffer << ' ' << x[2];
-      buffer << ' ' << x[3];
-      buffer << ' ' << x[4] << '\n';
-      x += 5;
-    }
-
-    if ( partial ) {
-
-      buffer << "   ";
-      while ( partial-- ) {
-
-        buffer << ' ' << *x;
-        ++x;
-      }
-      buffer << '\n';
-    }
-
-    for ( auto c : buffer.str() ) { *iter++ = c; }
-  };
+  #include "NDItk/base/IntegerListRecord/src/write.hpp"
 
 public:
 
@@ -93,22 +55,7 @@ public:
   using Parent::empty;
   using Parent::print;
 
-  /**
-   *  @brief Read the record data
-   *
-   *  @param[in] iter   the current position in the input
-   */
-  template< typename Iterator >
-  void read( Iterator& iter, const Iterator& end, std::size_t size ) {
-
-    std::vector< int > data;
-    data.reserve( size );
-    while ( size-- ) {
-
-      data.push_back( njoy::tools::disco::FreeFormatInteger::read< int >( iter, end ) );
-    }
-    this->data() = data;
-  };
+  #include "NDItk/base/IntegerListRecord/src/read.hpp"
 };
 
 } // base namespace
